tests/test_image_processor.py
"""
Unit tests for SpriteForge image processor

Run with: pytest tests/
"""

import pytest
from pathlib import Path
import numpy as np
from PIL import Image as PILImage

# Note: Uncomment when you want to add tests
# from spriteforge.image_processor import ImageProcessor


class TestImageProcessor:
    """Test cases for ImageProcessor class"""
    
    def test_placeholder(self):
        """Placeholder test - implement actual tests here"""
        assert True
    
    # Example test structure:
    # 
    # @pytest.fixture
    # def sample_image(self, tmp_path):
    #     """Create a sample test image"""
    #     img = PILImage.new('RGB', (100, 100), color='red')
    #     img_path = tmp_path / "test.png"
    #     img.save(img_path)
    #     return str(img_path)
    # 
    # def test_load_image(self, sample_image):
    #     """Test image loading"""
    #     processor = ImageProcessor(sample_image)
    #     assert processor.image is not None
    #     assert processor.get_image_size() == (100, 100)
    # 
    # def test_get_selection(self, sample_image):
    #     """Test region selection"""
    #     processor = ImageProcessor(sample_image)
    #     selection = (10, 10, 50, 50)
    #     sprite = processor.get_selection_image(selection)
    #     assert sprite is not None
    #     assert sprite.size == (50, 50)


if __name__ == '__main__':
    pytest.main([__file__, '-v'])
